requirements.txt
yt-dlp
ffmpeg-python
langchain_community
rank_bm25
faiss-cpu
langchain_huggingface
langchain_openai
langgraph
streamlit
git+https://github.com/openai/whisper.git
python-dotenv